Incomplete, uploaded for deliverable 2
Card, Deck, and PlayerHand structs are complete. Card and Deck structs allow for a deck to be created, shuffled, dealt, and displayed. PlayerHand struct holds a player's 7 cards, and detects the best 5-card hand out those 7 cards at the end of a round. Overloaded operators allow for comparison. 
All subsequent structs, classes, and functions are still in progress. They are all being modified as necessary as I work on the functions that will allow the poker game to run, mostly concerning processing the bets. These functions are kind of a mess right now because I'm completing them in a scattered manner, but will be functional and allow for the game to run soon.